Fixed inventory and price problems

+++ service/src/main/kotlin/com/example/videostore/cart/dto/SalesOrderDTO.kt
@@ -1,6 +1,7 @@
 package com.example.videostore.cart.dto
 
 import com.example.videostore.user.User
+import java.math.BigDecimal
 import java.time.LocalDateTime
 import java.util.*
 
@@ -11,6 +12,7 @@ data class SalesOrderDTO(
     val updatedOn: LocalDateTime,
     val salesOrderLines: MutableList<SalesOrderLineDTO>
 ) {
-    val totalPrice: Double =
-        salesOrderLines.map { it.price }.let { if (it.isNotEmpty()) it.reduce { acc, price -> price + acc } else 0.0 }
+    val totalPrice: BigDecimal =
+        (salesOrderLines.map { it.price }
+            .let { if (it.isNotEmpty()) it.reduce { acc, price -> price + acc } else BigDecimal(0.0) })
 }
